#14 add requireAuth
