+++ src/components/App/App.tsx
@@ -1,4 +1,4 @@
-import { Router } from '@/router/Router';
+import { Router } from '@/router/router';
 import { withRouter } from '@/components/App/providers/withRouter';
 
 const App = () => <Router />;
